core: move storage-related interop code into the storage package
