+++ pkg/core/interop/storage/find.go
@@ -1,6 +1,10 @@
 package storage
 
 import (
+	"context"
+	"fmt"
+
+	"github.com/nspcc-dev/neo-go/pkg/core/interop"
 	"github.com/nspcc-dev/neo-go/pkg/core/storage"
 	"github.com/nspcc-dev/neo-go/pkg/util/slice"
 	"github.com/nspcc-dev/neo-go/pkg/vm/stackitem"
@@ -79,4 +83,46 @@ func (s *Iterator) Value() stackitem.Item {
 		stackitem.NewByteArray(key),
 		value,
 	})
+}
+
+// Find finds stored key-value pair.
+func Find(ic *interop.Context) error {
+	stcInterface := ic.VM.Estack().Pop().Value()
+	stc, ok := stcInterface.(*Context)
+	if !ok {
+		return fmt.Errorf("%T is not a storage,Context", stcInterface)
+	}
+	prefix := ic.VM.Estack().Pop().Bytes()
+	opts := ic.VM.Estack().Pop().BigInt().Int64()
+	if opts&^FindAll != 0 {
+		return fmt.Errorf("%w: unknown flag", errFindInvalidOptions)
+	}
+	if opts&FindKeysOnly != 0 &&
+		opts&(FindDeserialize|FindPick0|FindPick1) != 0 {
+		return fmt.Errorf("%w KeysOnly conflicts with other options", errFindInvalidOptions)
+	}
+	if opts&FindValuesOnly != 0 &&
+		opts&(FindKeysOnly|FindRemovePrefix) != 0 {
+		return fmt.Errorf("%w: KeysOnly conflicts with ValuesOnly", errFindInvalidOptions)
+	}
+	if opts&FindPick0 != 0 && opts&FindPick1 != 0 {
+		return fmt.Errorf("%w: Pick0 conflicts with Pick1", errFindInvalidOptions)
+	}
+	if opts&FindDeserialize == 0 && (opts&FindPick0 != 0 || opts&FindPick1 != 0) {
+		return fmt.Errorf("%w: PickN is specified without Deserialize", errFindInvalidOptions)
+	}
+	ctx, cancel := context.WithCancel(context.Background())
+	seekres := ic.DAO.SeekAsync(ctx, stc.ID, storage.SeekRange{Prefix: prefix})
+	item := NewIterator(seekres, prefix, opts)
+	ic.VM.Estack().PushItem(stackitem.NewInterop(item))
+	ic.RegisterCancelFunc(func() {
+		cancel()
+		// Underlying persistent store is likely to be a private MemCachedStore. Thus,
+		// to avoid concurrent map iteration and map write we need to wait until internal
+		// seek goroutine is finished, because it can access underlying persistent store.
+		for range seekres {
+		}
+	})
+
+	return nil
 }
